Desktop and ODD Toast reunification and reconciliation [RAUT-379]
